# Conflicts:
#	dev/src/components/Menu.vue
#	views/css/buckaroo3.vue.css
#	views/js/buckaroo.vue.js

+++ upgrade/upgrade-3.4.php
@@ -14,6 +14,10 @@
  *  @copyright Copyright (c) Buckaroo B.V.
  *  @license   http://opensource.org/licenses/afl-3.0.php  Academic Free License (AFL 3.0)
  */
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
 // Remove old payment methods
 // empayment
 Configuration::deleteByName('BUCKAROO_EMPAYMENT_ENABLED');
